Moved REST API endpoints for addresses of the node wallet to '/wallet'. Removed REST API endpoints for exporting the seed.
