+++ src/main/scala/com/ltonetwork/api/http/AddressApiRoute.scala
@@ -1,8 +1,7 @@
 package com.ltonetwork.api.http
 
-import java.nio.charset.StandardCharsets
-
 import akka.http.scaladsl.marshalling.ToResponseMarshallable
+import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.server.Route
 import com.ltonetwork.account.{Address, PublicKeyAccount}
 import com.ltonetwork.consensus.GeneratingBalanceProvider
@@ -19,6 +18,7 @@ import com.ltonetwork.utx.UtxPool
 import com.ltonetwork.wallet.Wallet
 import io.netty.channel.group.ChannelGroup
 import io.swagger.annotations._
+
 import javax.ws.rs.Path
 import play.api.libs.json._
 
@@ -39,14 +39,12 @@ case class AddressApiRoute(settings: RestAPISettings,
 
   import AddressApiRoute._
 
-  val MaxAddressesPerRequest = 1000
-
   override lazy val route =
     pathPrefix("addresses") {
-      validate ~ seed ~ balanceWithConfirmations ~ balanceDetails ~ balanceHistory ~ balance ~
-        balanceWithConfirmations ~ verify ~ sign ~ deleteAddress ~ verifyText ~ signText ~ seq ~ publicKey ~
+      validate ~ balanceWithConfirmations ~ balanceDetails ~ balanceHistory ~ balance ~
+        balanceWithConfirmations ~ verify ~ verifyText ~ publicKey ~
         effectiveBalance ~ effectiveBalanceWithConfirmations ~ scriptInfo
-    } ~ root ~ create
+    } ~ root
 
   @Path("/scriptInfo/{address}")
   @ApiOperation(value = "Details for account", notes = "Account's script", httpMethod = "GET")
@@ -61,66 +59,6 @@ case class AddressApiRoute(settings: RestAPISettings,
         .flatMap(addressScriptInfoJson)
         .map(ToResponseMarshallable(_))
     )
-  }
-
-  @Path("/{address}")
-  @ApiOperation(value = "Delete", notes = "Remove the account with address {address} from the wallet", httpMethod = "DELETE")
-  @ApiImplicitParams(
-    Array(
-      new ApiImplicitParam(name = "address", value = "Address", required = true, dataType = "string", paramType = "path")
-    ))
-  def deleteAddress: Route = path(Segment) { address =>
-    (delete & withAuth) {
-      if (Address.fromString(address).isLeft) {
-        complete(InvalidAddress)
-      } else {
-        val deleted = wallet.findPrivateKey(address).flatMap(account => wallet.deleteAccount(account))
-        deleted match {
-          case Right(d) => complete(Json.obj("deleted" -> JsBoolean(d)))
-          case Left(e)  => complete(e)
-        }
-      }
-    }
-  }
-
-  @Path("/sign/{address}")
-  @ApiOperation(value = "Sign", notes = "Sign a message with a private key associated with {address}", httpMethod = "POST")
-  @ApiImplicitParams(
-    Array(
-      new ApiImplicitParam(name = "message", value = "Message to sign as a plain string", required = true, paramType = "body", dataType = "string"),
-      new ApiImplicitParam(name = "address", value = "Address", required = true, dataType = "string", paramType = "path")
-    ))
-  @ApiResponses(
-    Array(
-      new ApiResponse(
-        code = 200,
-        message =
-          "Json with error or json like {\"message\": \"Base58-encoded\",\"publickey\": \"Base58-encoded\", \"signature\": \"Base58-encoded\"}"
-      )
-    ))
-  def sign: Route = {
-    path("sign" / Segment) { address =>
-      signPath(address, encode = true)
-    }
-  }
-
-  @Path("/signText/{address}")
-  @ApiOperation(value = "Sign", notes = "Sign a message with a private key associated with {address}", httpMethod = "POST")
-  @ApiImplicitParams(
-    Array(
-      new ApiImplicitParam(name = "message", value = "Message to sign as a plain string", required = true, paramType = "body", dataType = "string"),
-      new ApiImplicitParam(name = "address", value = "Address", required = true, dataType = "string", paramType = "path")
-    ))
-  @ApiResponses(
-    Array(
-      new ApiResponse(
-        code = 200,
-        message = "Json with error or json like {\"message\": \"plain text\",\"publickey\": \"Base58-encoded\", \"signature\": \"Base58-encoded\"}")
-    ))
-  def signText: Route = {
-    path("signText" / Segment) { address =>
-      signPath(address, encode = false)
-    }
   }
 
   @Path("/verify/{address}")
@@ -244,21 +182,6 @@ case class AddressApiRoute(settings: RestAPISettings,
     }
   }
 
-  @Path("/seed/{address}")
-  @ApiOperation(value = "Seed", notes = "Export seed value for the {address}", httpMethod = "GET")
-  @ApiImplicitParams(
-    Array(
-      new ApiImplicitParam(name = "address", value = "Address", required = true, dataType = "string", paramType = "path")
-    ))
-  def seed: Route = {
-    (path("seed" / Segment) & get & withAuth) { address =>
-      complete(for {
-        pk   <- wallet.findPrivateKey(address)
-        seed <- wallet.exportAccountSeed(pk)
-      } yield Json.obj("address" -> address, "seed" -> Base58.encode(seed)))
-    }
-  }
-
   @Path("/validate/{address}")
   @ApiOperation(value = "Validate", notes = "Check whether address {address} is valid or not", httpMethod = "GET")
   @ApiImplicitParams(
@@ -272,39 +195,7 @@ case class AddressApiRoute(settings: RestAPISettings,
   @Path("/")
   @ApiOperation(value = "Addresses", notes = "Get wallet accounts addresses", httpMethod = "GET")
   def root: Route = (path("addresses") & get) {
-    val accounts = wallet.privateKeyAccounts
-    val json     = JsArray(accounts.map(a => JsString(a.address)))
-    complete(json)
-  }
-
-  @Path("/seq/{from}/{to}")
-  @ApiOperation(value = "Seq", notes = "Get wallet accounts addresses", httpMethod = "GET")
-  @ApiImplicitParams(
-    Array(
-      new ApiImplicitParam(name = "from", value = "Start address", required = true, dataType = "integer", paramType = "path"),
-      new ApiImplicitParam(name = "to", value = "address", required = true, dataType = "integer", paramType = "path")
-    ))
-  def seq: Route = {
-    (path("seq" / IntNumber / IntNumber) & get) {
-      case (start, end) =>
-        if (start >= 0 && end >= 0 && start - end < MaxAddressesPerRequest) {
-          val json = JsArray(
-            wallet.privateKeyAccounts.map(a => JsString(a.address)).slice(start, end)
-          )
-
-          complete(json)
-        } else complete(TooBigArrayAllocation)
-    }
-  }
-
-  @Path("/")
-  @ApiOperation(value = "Create", notes = "Create a new account in the wallet(if it exists)", httpMethod = "POST")
-  def create: Route = (path("addresses") & post & withAuth) {
-    wallet.generateNewAccount() match {
-      case Right(Some(pka)) => complete(Json.obj("address" -> pka.address))
-      case Right(None)      => complete(Unknown)
-      case Left(e)          => complete(e)
-    }
+    redirect("/wallet/addresses", StatusCodes.PermanentRedirect)
   }
 
   private def balanceJson(address: String, confirmations: Int): ToResponseMarshallable = {
@@ -397,20 +288,6 @@ case class AddressApiRoute(settings: RestAPISettings,
       value <- blockchain.accountData(addr, key).toRight(DataKeyNotExists)
     } yield value
     ToResponseMarshallable(result)
-  }
-
-  private def signPath(address: String, encode: Boolean) = (post & entity(as[String])) { message =>
-    withAuth {
-      val res = wallet
-        .findPrivateKey(address)
-        .map(pk => {
-          val messageBytes = message.getBytes(StandardCharsets.UTF_8)
-          val signature    = crypto.sign(pk, messageBytes)
-          val msg          = if (encode) Base58.encode(messageBytes) else message
-          Signed(msg, Base58.encode(pk.publicKey), Base58.encode(signature))
-        })
-      complete(res)
-    }
   }
 
   private def verifyPath(address: String, decode: Boolean) = withAuth {
